Move the request log to be public
